move components to special directory
