+++ static/public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Title</title>
+    <title>Fire Marshal</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
 </head>
 <body>
