+++ static/src/components/landing/index.js
