+++ static/src/components/landing/landing.scss
