+++ static/src/components/navbar/navbar.scss
